Refactor common.lua into three files
This puts generic_... functions into generic.lua, hstore and other
functions which convert to standard PostgreSQL types in util.lua,
and leaves ClearTables specific stuff like height and maxspeed in
common.lua
